changes from pr comments
