+++ _layouts/base.html
@@ -19,8 +19,7 @@
 <body>
   {{ content }}
 
-  
-  <script src="https://unpkg.com/typer-js"></script>
+
   <script type="text/javascript" src="https://code.jquery.com/jquery-3.1.1.min.js"></script>
   <script type="text/javascript" src="/js/main.js"></script>
