+++ about.html
@@ -11,12 +11,7 @@ permalink: /about/
 
     <div class="project-section">
       <div class="about__container">
-
-        <!-- <div class="about__text--mask">
-          <div class="about__text--hover">
-            <h1 class="about__text">hi, i’m Julie:<br>an interaction designer who loves the problem solving that goes on behind screens.</h1>
-          </div>
-        </div> -->
+        
         <div class="about__text--before before-load">
           <h1 class="about__text">hi, i’m Julie,</h1>
           <h1 class="about__text about__text--fact"></h1>
